create-adb-connections/requirements.txt
python-dotenv
requests
azure-identity
cryptography